Add Event Deletion and Update Event show view
Create destroy method for EventsController and update Event show view.

The destroy method takes the given id parameter and searches for the
record in the current_user's hosted_events association. If it isn't
found, the resulting error is rescued and the user is redirected to the
show event page with a flash alert.

Show view now shows a delete button on the event's show page if the
current_user is the host of the event. A confirmation box is displayed
before sending the delete request.